Refactoring to a loader format
